Moved util functions to seperate dir
