+++ bin/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './rate-limitter';
 export * from './pathFromUrl';
+export * from './auth.service.utils';
